Divide new component ProductStars from ProductBox
