+++ src/components/common/ProductBox/ProductBox.js
@@ -3,16 +3,12 @@ import PropTypes from 'prop-types';
 import clsx from 'clsx';
 import styles from './ProductBox.module.scss';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import {
-  faStar,
-  faExchangeAlt,
-  faShoppingBasket,
-} from '@fortawesome/free-solid-svg-icons';
-import { faStar as farStar, faHeart } from '@fortawesome/free-regular-svg-icons';
+import { faExchangeAlt, faShoppingBasket } from '@fortawesome/free-solid-svg-icons';
+import { faHeart } from '@fortawesome/free-regular-svg-icons';
 import Button from '../Button/Button';
 import { useDispatch, useSelector } from 'react-redux';
 import { changeCompare } from '../../../redux/productsRedux';
-import { updateUserStars } from '../../../redux/productsRedux';
+import ProductStars from '../../features/ProductStars/ProductStars';
 
 const ProductBox = ({
   id,
@@ -20,6 +16,7 @@ const ProductBox = ({
   price,
   promo,
   stars,
+  userStars,
   isFavorite,
   isCompare,
   image,
@@ -42,16 +39,6 @@ const ProductBox = ({
     dispatch(changeCompare(id));
   };
 
-  const updateStars = (e, userStars, id) => {
-    e.preventDefault();
-    const changeStars = {
-      id,
-      userStars,
-    };
-    dispatch(updateUserStars(changeStars));
-    console.log('updated stars', changeStars);
-  };
-
   return (
     <div className={styles.root}>
       <div className={styles.photo} style={{ backgroundImage: `url(${image})` }}>
@@ -65,22 +52,7 @@ const ProductBox = ({
       </div>
       <div className={styles.content}>
         <h5>{name}</h5>
-        <div className={styles.stars}>
-          {[1, 2, 3, 4, 5].map(i => (
-            <Button
-              key={i}
-              onClick={e => {
-                updateStars(e, i, id);
-              }}
-            >
-              {i <= stars ? (
-                <FontAwesomeIcon icon={faStar}>{i} stars</FontAwesomeIcon>
-              ) : (
-                <FontAwesomeIcon icon={farStar}>{i} stars</FontAwesomeIcon>
-              )}
-            </Button>
-          ))}
-        </div>
+        <ProductStars stars={stars} userStars={userStars} id={id}></ProductStars>
       </div>
       <div className={styles.line}></div>
       <div className={styles.actions}>
@@ -114,6 +86,7 @@ ProductBox.propTypes = {
   price: PropTypes.number,
   promo: PropTypes.string,
   stars: PropTypes.number,
+  userStars: PropTypes.number,
   isFavorite: PropTypes.bool,
   isCompare: PropTypes.bool,
   image: PropTypes.string,
